refactor(modal-shell): extract frame focus helpers
